+++ CesiumITwinClient/include/CesiumITwinClient/Connection.h
@@ -17,7 +17,7 @@
 #include <CesiumITwinClient/PagedList.h>
 #include <CesiumUtility/Result.h>
 #include <CesiumUtility/Uri.h>
-#include <CesiumVectorData/GeoJsonObject.h>
+#include <CesiumVectorData/GeoJsonObjectTypes.h>
 
 #include <memory>
 #include <optional>
@@ -198,7 +198,7 @@ public:
    * @param limit The maximum number of items per page, between 1 and 10,000.
    */
   CesiumAsync::Future<
-      CesiumUtility::Result<PagedList<CesiumVectorData::GeoJsonObject>>>
+      CesiumUtility::Result<PagedList<CesiumVectorData::GeoJsonFeature>>>
   geospatialFeatures(
       const std::string& iTwinId,
       const std::string& collectionId,
@@ -224,12 +224,12 @@ public:
   Connection(
       const CesiumAsync::AsyncSystem& asyncSystem,
       const std::shared_ptr<CesiumAsync::IAssetAccessor>& pAssetAccessor,
-      const AuthenticationToken& authToken,
+      const AuthenticationToken& authenticationToken,
       const std::optional<std::string>& refreshToken,
       const CesiumClientCommon::OAuth2ClientOptions& clientOptions)
       : _asyncSystem(asyncSystem),
         _pAssetAccessor(pAssetAccessor),
-        _authToken(authToken),
+        _authenticationToken(authenticationToken),
         _refreshToken(refreshToken),
         _clientOptions(clientOptions) {}
 
@@ -237,29 +237,31 @@ public:
    * @brief Returns the \ref AuthenticationToken object representing the parsed
    * JWT access or share token.
    */
-  const AuthenticationToken& getAuthToken() const { return _authToken; }
+  const AuthenticationToken& getAuthenticationToken() const {
+    return this->_authenticationToken;
+  }
 
   /**
    * @brief Sets the access or share token that will be used for API calls.
    *
    * @param authToken The new auth token.
    */
-  void setAuthToken(const AuthenticationToken& authToken) {
-    _authToken = authToken;
+  void setAuthenticationToken(const AuthenticationToken& authToken) {
+    this->_authenticationToken = authToken;
   }
 
   /**
    * @brief Returns the refresh token used to obtain new access tokens, if any.
    */
   const std::optional<std::string>& getRefreshToken() const {
-    return _refreshToken;
+    return this->_refreshToken;
   }
 
   /**
    * @brief Sets the refresh token used to obtain new access tokens, if any.
    */
   void setRefreshToken(const std::optional<std::string>& refreshToken) {
-    _refreshToken = refreshToken;
+    this->_refreshToken = refreshToken;
   }
 
 private:
@@ -272,14 +274,14 @@ private:
   CesiumAsync::Future<CesiumUtility::Result<PagedList<ITwinRealityData>>>
   listITwinRealityData(const std::string& url);
   CesiumAsync::Future<
-      CesiumUtility::Result<PagedList<CesiumVectorData::GeoJsonObject>>>
+      CesiumUtility::Result<PagedList<CesiumVectorData::GeoJsonFeature>>>
   listGeospatialFeatures(const std::string& url);
 
   CesiumAsync::Future<CesiumUtility::Result<std::string>> ensureValidToken();
 
   CesiumAsync::AsyncSystem _asyncSystem;
   std::shared_ptr<CesiumAsync::IAssetAccessor> _pAssetAccessor;
-  AuthenticationToken _authToken;
+  AuthenticationToken _authenticationToken;
   std::optional<std::string> _refreshToken;
   CesiumClientCommon::OAuth2ClientOptions _clientOptions;
 };