+++ CesiumITwinClient/include/CesiumITwinClient/IModel.h
@@ -50,6 +50,6 @@ struct IModel {
    * @brief The maximum rectangular area on the Earth which encloses the iModel.
    */
   CesiumGeospatial::GlobeRectangle extent =
-      CesiumGeospatial::GlobeRectangle(0, 0, 0, 0);
+      CesiumGeospatial::GlobeRectangle::EMPTY;
 };
 }; // namespace CesiumITwinClient